python_logs/requirements.txt
paramiko>=2.7.2
requests>=2.25.1
pandas>=1.3.0
matplotlib>=3.4.2
seaborn>=0.11.1
pyyaml>=5.4.1
psutil>=5.8.0
schedule>=1.1.0
click>=8.0.1
jinja2>=3.0.1
sqlalchemy>=1.4.22
pymongo>=3.12.0
elasticsearch>=7.14.0
redis>=3.5.3